pyproject.toml
[project]
name = "transit_reader"
version = "0.2.0"
description = "transit_reader using crewAI"
authors = [{ name = "Benjamin Jasper", email = "ben.j.jasper@gmail.com" }]
requires-python = ">=3.12.9,<3.13"
dependencies = [
    "crewai[tools]>=0.126.0,<1.0.0",
    "html2text",
    "python-dotenv>=1.1.0",
    "trafilatura>=2.0.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.19.0",
    "botocore>=1.38.32",
    "markdown>=3.7",
    "immanuel>=1.5.0",
    "kerykeion==4.26.2",
    "weasyprint>=65.1",
    "boto3>=1.37.36",
    "md2pdf>=1.0.1",
    "selenium>=4.33.0",
    "ipykernel>=6.29.5",
    "googlemaps>=4.10.0",
    "linkup-sdk>=0.2.6",
]

[project.scripts]
kickoff = "transit_reader.main:kickoff"
plot = "transit_reader.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
